COMMIT-2

+++ src/main/java/com/algomau/bank/domain/Account.java
@@ -10,7 +10,6 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -29,7 +28,7 @@ public class Account {
     private UUID id;
 
     @Column(name = "balance")
-    private BigDecimal balance;
+    private double balance;
 
     @Column(name = "type")
     @Enumerated(EnumType.STRING)
